usecases merged
