+++ backend/AAiT-backend-group-13/usecase/user/command/login/command.go
@@ -1,0 +1,14 @@
+package logincommand
+
+
+type LoginCommand struct {
+	username string
+	password string
+}
+
+func NewLoginCommand(username string, password string) LoginCommand {
+	return LoginCommand{
+		username: username,
+		password: password,
+	}
+}
