+++ backend/AAiT-backend-group-13/usecase/user/command/signup/command.go
@@ -1,4 +1,4 @@
-package usercommand
+package signcommand
 
 type SignUpCommand struct {
 	firstName string
@@ -17,18 +17,3 @@ func NewSignUpCommand(username string, firstName string, lastName string, email 
 		password:  password,
 	}
 }
-
-type LoginCommand struct {
-	username string
-	password string
-}
-
-func NewLoginCommand(username string, password string) LoginCommand {
-	return LoginCommand{
-		username: username,
-		password: password,
-	}
-}
-
-
-
