+++ backend/AAiT-backend-group-13/usecase/user/result/result.go
