--- backend/AAiT-backend-group-13/usecases_sof/utils/command/command.go
@@ -1,5 +1,0 @@
-package icommand
-
-type Ihandler [Command any, Result any] interface{
-	Handle(command Command) (Result , error)
-}
