--- backend/AAiT-backend-group-13/usecases_sof/utils/i_email/email_service.go
@@ -1,5 +1,0 @@
-package iemail
-
-type Service interface {
-  Send(mail *Mail) error
-}
