--- backend/AAiT-backend-group-13/usecases_sof/utils/query/query.go
@@ -1,6 +1,0 @@
-package iquery 
-
-
-type Ihandler[Query any, Result any] interface {
-	Handle(query Query) ( Result, error)
-}
